Watch App 세팅중...

+++ lib/thread/view/video_editor_screen.dart
@@ -76,7 +76,8 @@ class _VideoEditScreenState extends ConsumerState<VideoEditorScreen> {
     await MediaUtils.runFFmpegCommand(
       await config.getExecuteConfig(),
       onProgress: (stats) {
-        _exportingProgress.value = config.getFFmpegProgress(stats.getTime());
+        _exportingProgress.value =
+            config.getFFmpegProgress(stats.getTime().toInt());
       },
       onError: (e, s) {
         debugPrint('video edit error ===> $e');
